Refactor API structure and implement OpenAI-compatible endpoints
- Moved API routes to separate modules for better organization and maintainability.
- Added OpenAI-compatible models endpoint and corresponding tests.
- Refactored fact processing methods to enhance context awareness and confidence estimation.
- Updated tests to use pytest for improved readability and maintainability.
- Created utility functions for token estimation and context extraction.
- Added comprehensive API documentation in README.md.
- Updated .gitignore to exclude unnecessary files.
